requirements.txt
Django>=4.0
djangorestframework>=3.12
django-cors-headers>=3.10
djangorestframework-simplejwt>=4.8
psycopg2>=2.9
django-filter
